+++ ServerData.cs
@@ -145,5 +145,5 @@ public record RouteInfo(string Route, string Method, bool IsLegacy = false);
 
 public static class APIVersion
 {
-    public const string Version = "1.7.0";
+    public const string Version = "1.8.0";
 }